# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Benefits Enrollment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /integrations:
    get:
      operationId: listIntegrations
      tags: [integrations]
      summary: List configured payroll integrations
      responses:
        "200":
          description: List of integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Integration"
    post:
      operationId: createIntegration
      tags: [integrations]
      summary: Configure a payroll integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationInput"
      responses:
        "201":
          description: Created integration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
  /integrations/{id}:
    get:
      operationId: getIntegration
      tags: [integrations]
      summary: Get integration by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Integration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
    patch:
      operationId: updateIntegration
      tags: [integrations]
      summary: Update integration settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationUpdate"
      responses:
        "200":
          description: Updated integration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
    delete:
      operationId: deleteIntegration
      tags: [integrations]
      summary: Delete integration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /integrations/{id}/sync:
    post:
      operationId: syncIntegration
      tags: [integrations]
      summary: Sync employees from payroll integration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
components:
  schemas:
...
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
        message:
          type: string
      required:
        - imported
        - skipped
        - errors
        - message
...
    Integration:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        provider:
          type: string
        status:
          type: string
        lastSyncAt:
          type: ["string", "null"]
        syncedEmployees:
          type: integer
        apiEndpoint:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - provider
        - status
        - syncedEmployees
        - createdAt
    IntegrationInput:
      type: object
      properties:
        name:
          type: string
        provider:
          type: string
        apiKey:
          type: string
        apiEndpoint:
          type: string
        notes:
          type: string
      required:
        - name
        - provider
    IntegrationUpdate:
      type: object
      properties:
        name:
          type: string
        apiKey:
          type: string
        apiEndpoint:
          type: string
        notes:
          type: string
        status:
          type: string